+++ src/WorkflowValidation/Tools/WorkflowTools.cs
@@ -39,7 +39,7 @@ namespace WorkflowValidation.Tools
         /// </summary>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflow StartWith(Action step)
+        public static IWorkflowStep StartWith(Action step)
         {
             return WorkflowValidation.Workflow.StartWith(step);
         }
@@ -51,7 +51,7 @@ namespace WorkflowValidation.Tools
         /// <param name="description"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflow StartWith(string description, Action step)
+        public static IWorkflowStep StartWith(string description, Action step)
         {
             return WorkflowValidation.Workflow.StartWith(description, step);
         }
@@ -62,7 +62,7 @@ namespace WorkflowValidation.Tools
         /// </summary>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflow StartWith(Action<WorkflowContext> step)
+        public static IWorkflowStep StartWith(Action<WorkflowContext> step)
         {
             return WorkflowValidation.Workflow.StartWith(step);
         }
@@ -73,16 +73,31 @@ namespace WorkflowValidation.Tools
         /// <param name="description"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflow StartWith(string description, Action<WorkflowContext> step)
+        public static IWorkflowStep StartWith(string description, Action<WorkflowContext> step)
         {
             return WorkflowValidation.Workflow.StartWith(description, step);
         }
 
+        /// <summary>
+        /// Setup for a new workflow
+        /// </summary>
+        /// <param name="setup"></param>
+        /// <returns></returns>
         public static IWorkflowBuilder SetupWorkflow(Action<IWorkflowSetupBuilder> setup)
         {
             var builder = new WorkflowBuilder();
             return builder.SetupWorkflow(setup);
         }
         
+        /// <summary>
+        /// Setup for a new workflow
+        /// </summary>
+        /// <param name="description"></param>
+        /// <returns></returns>
+        public static IWorkflowBuilder SetupWorkflow(string description)
+        {
+            var builder = new WorkflowBuilder();
+            return builder.SetupWorkflow(description);
+        }
     }
 }