+++ src/WorkflowValidation/WorkflowExtensions.cs
@@ -14,7 +14,7 @@ namespace WorkflowValidation
         /// <param name="workflow"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep Then(this IWorkflow workflow, Action step)
+        public static IWorkflowStep Then(this IWorkflowStep workflow, Action step)
         {
             workflow.SetStep(new Step(step));
             return new Workflow(workflow);
@@ -27,7 +27,7 @@ namespace WorkflowValidation
         /// <param name="name"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep Then(this IWorkflow workflow,string name, Action step)
+        public static IWorkflowStep Then(this IWorkflowStep workflow,string name, Action step)
         {
             workflow.SetStep(new Step(step)
                 .SetName(name)
@@ -42,7 +42,7 @@ namespace WorkflowValidation
         /// <param name="workflow"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep Then(this IWorkflow workflow, Action<WorkflowContext> step)
+        public static IWorkflowStep Then(this IWorkflowStep workflow, Action<WorkflowContext> step)
         {
             workflow.SetStep(new Step(step));
             return new Workflow(workflow);
@@ -55,7 +55,7 @@ namespace WorkflowValidation
         /// <param name="message"></param>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep Then(this IWorkflow workflow, string message, Action<WorkflowContext> step)
+        public static IWorkflowStep Then(this IWorkflowStep workflow, string message, Action<WorkflowContext> step)
         {
             workflow.SetStep(new Step(step)
                 .SetName(message)
@@ -70,6 +70,7 @@ namespace WorkflowValidation
         /// <param name="workflow"></param>
         /// <param name="ensure"></param>
         /// <returns></returns>
+        [AssertionMethod]
         public static IWorkflowStep Verify(this IWorkflowStep workflow, Action<AssertionProvider> ensure)
         {
             workflow.SetStep(new AssertionStep(c =>
@@ -81,20 +82,30 @@ namespace WorkflowValidation
             return new Workflow(workflow);
         }
 
+        [AssertionMethod]
+        public static IWorkflowStep Verify(this IWorkflowStep workflow, string name, Func<bool> assert)
+        {
+            workflow.SetStep(new AssertionStep(assert)
+                .SetName(name)
+            );
+
+            return new Workflow(workflow);
+        }
+
         /// <summary>
         /// Waits on the thread for a given amount of milliseconds
         /// </summary>
         /// <param name="workflow"></param>
         /// <param name="milliseconds"></param>
         /// <returns></returns>
-        public static IWorkflow Wait(this IWorkflow workflow, int milliseconds)
+        public static IWorkflow Wait(this IWorkflowStep workflow, int milliseconds)
         {
             workflow.SetStep(new Step(() =>
                 {
-                    //Thread.Sleep(milliseconds);
                     System.Threading.Tasks.Task.Delay(milliseconds).Wait();
                 })
-                .SetName($"Wait for {milliseconds} ms"));
+                .SetName($"Wait for {milliseconds} ms")
+            );
 
             return workflow;
         }